refactor: add ids to sections
