+++ src/components/home/blog-overview.astro
@@ -4,8 +4,8 @@ import { Button } from '@/components/ui/button';
 import { Card, CardHeader, CardContent } from '@/components/ui/card';
 import { getCollection } from 'astro:content';
 import { Image } from 'astro:assets';
-import { Spacer } from '@/components/spacer';
-import FormattedDate from '@/components/formatted-date.astro';
+import { Spacer } from '@/components/ui/spacer';
+import FormattedDate from '@/components/blog/formatted-date.astro';
 
 const sortByDate = (a: any, b: any) =>
   new Date(b.data.date).valueOf() - new Date(a.data.date).valueOf();
@@ -23,6 +23,7 @@ const postImages = posts.map((post) => {
 ---
 
 <section
+  id="blog"
   class="mx-auto flex w-10/12 flex-col items-start py-12 lg:py-24 xl:w-10/12 2xl:w-[1280px]"
 >
   <h1 class="text-text-1 text-4xl">Read our blog</h1>
