+++ src/components/home/features.astro
@@ -9,7 +9,7 @@ import {
   CardFooter,
 } from '@/components/ui/card';
 import { Ellipse } from '@/components/ellipse';
-import { Spacer } from '@/components/spacer';
+import { Spacer } from '@/components/ui/spacer';
 import {
   ArrowRight,
   Lightning,
@@ -69,6 +69,7 @@ const features = [
 ---
 
 <section
+  id="features"
   class="mx-auto flex w-10/12 flex-col items-start py-12 lg:py-24 xl:w-10/12 2xl:w-[1280px]"
 >
   <h1 class="text-text-1 text-4xl">
