+++ src/components/home/for-developers.astro
@@ -8,12 +8,13 @@ import SocialProof from '@/assets/features/social-proof.png';
 ---
 
 <section
+  id="developers"
   class="mx-auto flex w-10/12 flex-col py-12 lg:py-24 xl:w-10/12 2xl:w-[1280px]"
 >
-  <div class="grid grid-cols-1 lg:grid-cols-2 gap-12">
+  <div class="grid grid-cols-1 gap-12 lg:grid-cols-2">
     <Image class="rounded-xl" src={QuickFix} alt="Fix bugs quickly" />
     <div class="flex flex-col justify-center gap-6 lg:items-start">
-      <div class="flex flex-row gap-4 items-center">
+      <div class="flex flex-row items-center gap-4">
         <Image height={36} src={SocialProof} alt="Social proof" />
         <span class="text-text-1 md:text-lg">2,000+ teams use us</span>
       </div>
@@ -23,19 +24,19 @@ import SocialProof from '@/assets/features/social-proof.png';
         waiting for full app releases.
       </span>
       <ul class="flex flex-col gap-4">
-        <li class="flex flex-row gap-2 items-start">
-          <Check className="size-6 text-accent-primary-1" /> Iterate quickly and
+        <li class="flex flex-row items-start gap-2">
+          <Check className="text-accent-primary-1 size-6" /> Iterate quickly and
           confidently
         </li>
-        <li class="flex flex-row gap-2 items-start">
-          <Check className="size-6 text-accent-primary-1" />Hotfix in seconds,
+        <li class="flex flex-row items-start gap-2">
+          <Check className="text-accent-primary-1 size-6" />Hotfix in seconds,
           not days
         </li>
-        <li class="flex flex-row gap-2 items-start">
-          <Check className="size-6 text-accent-primary-1" />No need for legacy
+        <li class="flex flex-row items-start gap-2">
+          <Check className="text-accent-primary-1 size-6" />No need for legacy
           code path maintenance
         </li>
-        <div class="flex w-full flex-row gap-4 mt-6 items-start">
+        <div class="mt-6 flex w-full flex-row items-start gap-4">
           <a href={config.consoleUrl}>
             <GradientOutlineButton className="font-light">
               Get started <ArrowRight className="size-5" weight="bold" />
