+++ src/components/home/for-teams.astro
@@ -7,9 +7,10 @@ import Releases from '@/assets/features/releases.png';
 ---
 
 <section
+  id="teams"
   class="mx-auto flex w-10/12 flex-col py-12 lg:py-24 xl:w-10/12 2xl:w-[1280px]"
 >
-  <div class="grid grid-cols-1 lg:grid-cols-2 gap-12">
+  <div class="grid grid-cols-1 gap-12 lg:grid-cols-2">
     <div class="flex flex-col justify-center gap-6 lg:items-start">
       <h2 class="text-4xl font-medium">For Teams</h2>
       <span class="text-text-2">
@@ -18,19 +19,19 @@ import Releases from '@/assets/features/releases.png';
         review cycle.
       </span>
       <ul class="flex flex-col gap-4">
-        <li class="flex flex-row gap-2 items-start">
-          <Check className="size-6 text-accent-primary-1" /> Respond to customer
+        <li class="flex flex-row items-start gap-2">
+          <Check className="text-accent-primary-1 size-6" /> Respond to customer
           issues and complaints quickly
         </li>
-        <li class="flex flex-row gap-2 items-start">
-          <Check className="size-6 text-accent-primary-1" />Easily phase out
+        <li class="flex flex-row items-start gap-2">
+          <Check className="text-accent-primary-1 size-6" />Easily phase out
           support for older app versions
         </li>
-        <li class="flex flex-row gap-2 items-start">
-          <Check className="size-6 text-accent-primary-1" />Fix customer
+        <li class="flex flex-row items-start gap-2">
+          <Check className="text-accent-primary-1 size-6" />Fix customer
           complaints in seconds, not days
         </li>
-        <div class="flex w-full flex-row gap-4 mt-6 items-start">
+        <div class="mt-6 flex w-full flex-row items-start gap-4">
           <a href={config.consoleUrl}>
             <GradientOutlineButton className="font-light">
               Get started <ArrowRight className="size-5" weight="bold" />
