+++ src/components/home/pricing-overview.astro
@@ -1,10 +1,11 @@
 ---
-import { PricingCalculator } from '@/components/pricing-calculator';
-import PricingPlans from '@/components/pricing-plans.astro';
-import { Spacer } from '@/components/spacer';
+import { PricingCalculator } from '@/components/pricing/pricing-calculator';
+import PricingPlans from '@/components/pricing/pricing-plans.astro';
+import { Spacer } from '@/components/ui/spacer';
 ---
 
 <section
+  id="pricing"
   class="mx-auto flex w-10/12 flex-col py-12 lg:py-24 xl:w-10/12 2xl:w-[1280px]"
 >
   <div class="flex flex-col items-center justify-center">
