+++ src/components/home/problem.astro
@@ -1,7 +1,7 @@
 ---
 import { Card, CardHeader, CardTitle, CardContent } from '@/components/ui/card';
 import { ClockCountdown, ChartLine, MoneyWavy } from '@phosphor-icons/react';
-import { Spacer } from '@/components/spacer';
+import { Spacer } from '@/components/ui/spacer';
 
 const problems = [
   {
@@ -26,6 +26,7 @@ const problems = [
 ---
 
 <section
+  id="problem"
   class="mx-auto flex w-10/12 flex-col py-12 lg:py-24 xl:w-10/12 2xl:w-[1280px]"
 >
   <div class="flex flex-col items-center justify-center">
