+++ src/components/home/resources.astro
@@ -1,7 +1,7 @@
 ---
 import { Card, CardHeader, CardTitle, CardContent } from '@/components/ui/card';
 import { Image } from 'astro:assets';
-import { Spacer } from '@/components/spacer';
+import { Spacer } from '@/components/ui/spacer';
 import config from '@/config';
 import Discord from '@/assets/resources/discord.png';
 import Shorebird from '@/assets/resources/shorebird.png';
@@ -25,6 +25,7 @@ const resources = [
 ---
 
 <section
+  id="resources"
   class="mx-auto flex w-11/12 flex-col items-center py-12 lg:py-24 xl:w-10/12 2xl:w-[1280px]"
 >
   <h1 class="text-text-1 text-4xl">Resources</h1>
