+++ src/components/logos/discord-logo.tsx
