+++ src/components/logos/logo-full.tsx
