+++ src/components/pricing/pricing-details.astro
@@ -9,10 +9,10 @@ import {
   TableRow,
 } from '@/components/ui/table';
 import { GradientOutlineButton } from '@/components/ui/button';
-import { PricingCalculator } from '@/components/pricing-calculator';
-import PricingPlans from '@/components/pricing-plans.astro';
-import { PricingFaq } from '@/components/pricing-faq';
-import { Spacer } from '@/components/spacer';
+import { PricingCalculator } from '@/components/pricing/pricing-calculator';
+import PricingPlans from '@/components/pricing/pricing-plans.astro';
+import { PricingFaq } from '@/components/pricing/pricing-faq';
+import { Spacer } from '@/components/ui/spacer';
 import { ArrowRight, Check, Minus } from '@phosphor-icons/react';
 
 interface Feature {
