+++ src/components/pricing/pricing-faq.tsx
@@ -1,4 +1,4 @@
-import { Markdown } from '@/components/markdown';
+import { Markdown } from '@/components/ui/markdown';
 import {
   Accordion,
   AccordionContent,
