+++ src/components/ui/footer.astro
@@ -1,11 +1,11 @@
 ---
 import { Button } from '@/components/ui/button';
-import { DiscordLogo } from '@/components/discord-logo';
-import { GitHubLogo } from '@/components/github-logo';
-import { LinkedInLogo } from '@/components/linkedin-logo';
-import { LogoFull } from '@/components/logo-full';
-import { Spacer } from '@/components/spacer';
-import { TwitterLogo } from '@/components/twitter-logo';
+import { DiscordLogo } from '@/components/logos/discord-logo';
+import { GitHubLogo } from '@/components/logos/github-logo';
+import { LinkedInLogo } from '@/components/logos/linkedin-logo';
+import { LogoFull } from '@/components/logos/logo-full';
+import { Spacer } from '@/components/ui/spacer';
+import { TwitterLogo } from '@/components/logos/twitter-logo';
 import config from '@/config';
 ---
