+++ src/layouts/blog.astro
@@ -1,10 +1,10 @@
 ---
 import '@/styles/global.css';
-import { BlogCallout } from '@/components/blog-callout';
+import { BlogCallout } from '@/components/blog/blog-callout';
 import { ClockCountdown } from '@phosphor-icons/react';
 import { Image } from 'astro:assets';
-import { Navbar } from '@/components/navbar';
-import { Spacer } from '@/components/spacer';
+import { Navbar } from '@/components/ui/navbar';
+import { Spacer } from '@/components/ui/spacer';
 import bryanHeadshot from '@/assets/blog/headshots/bryan-headshot.png';
 import ericHeadshot from '@/assets/blog/headshots/eric-headshot.jpeg';
 import felixHeadshot from '@/assets/blog/headshots/felix-headshot.jpeg';
