+++ src/layouts/markdown.astro
@@ -1,6 +1,6 @@
 ---
 import '@/styles/global.css';
-import { Navbar } from '@/components/navbar';
+import { Navbar } from '@/components/ui/navbar';
 import Footer from '@/components/footer.astro';
 
 const { frontmatter } = Astro.props;
