+++ src/layouts/success-story.astro
@@ -2,8 +2,8 @@
 import '@/styles/global.css';
 import { Image } from 'astro:assets';
 import { getCollection } from 'astro:content';
-import { BlogCallout } from '@/components/blog-callout';
-import { Navbar } from '@/components/navbar';
+import { BlogCallout } from '@/components/blog/blog-callout';
+import { Navbar } from '@/components/ui/navbar';
 import Footer from '@/components/footer.astro';
 import FormattedDate from '@/components/formatted-date.astro';
 import { ClockCountdown } from '@phosphor-icons/react';
