+++ src/pages/about.astro
@@ -3,12 +3,12 @@ import '@/styles/global.css';
 import { Button, GradientOutlineButton } from '@/components/ui/button';
 import { Card } from '@/components/ui/card';
 import { Ellipse } from '@/components/ellipse';
-import { FlutterLogo } from '@/components/flutter-logo';
+import { FlutterLogo } from '@/components/logos/flutter-logo';
 import { Image } from 'astro:assets';
-import { LinkedInLogo } from '@/components/linkedin-logo';
-import { Markdown } from '@/components/markdown';
-import { Navbar } from '@/components/navbar';
-import { Spacer } from '@/components/spacer';
+import { LinkedInLogo } from '@/components/logos/linkedin-logo';
+import { Markdown } from '@/components/ui/markdown';
+import { Navbar } from '@/components/ui/navbar';
+import { Spacer } from '@/components/ui/spacer';
 import Banner from '@/components/banner.astro';
 import bryanHeadshot from '@/assets/team/bryan-headshot.png';
 import config from '@/config';
