+++ src/pages/blog/index.astro
@@ -5,9 +5,9 @@ import { Button } from '@/components/ui/button';
 import { Card, CardHeader, CardContent } from '@/components/ui/card';
 import { getCollection } from 'astro:content';
 import { Image } from 'astro:assets';
-import { Navbar } from '@/components/navbar';
+import { Navbar } from '@/components/ui/navbar';
 import { ScrollToTopButton } from '@/components/scroll-to-top-button';
-import { Spacer } from '@/components/spacer';
+import { Spacer } from '@/components/ui/spacer';
 import Banner from '@/components/banner.astro';
 import config from '@/config';
 import Footer from '@/components/footer.astro';
