+++ src/pages/pricing.astro
@@ -1,8 +1,8 @@
 ---
 import MainLayout from '@/layouts/main.astro';
 import Banner from '@/components/banner.astro';
-import { Navbar } from '@/components/navbar';
-import PricingDetails from '@/components/pricing-details.astro';
+import { Navbar } from '@/components/ui/navbar';
+import PricingDetails from '@/components/pricing/pricing-details.astro';
 import Footer from '@/components/footer.astro';
 ---
