+++ src/pages/success-stories/index.astro
@@ -6,9 +6,9 @@ import { ArrowRight } from '@phosphor-icons/react';
 import { Card, CardHeader, CardContent } from '@/components/ui/card';
 import { getCollection } from 'astro:content';
 import { Image } from 'astro:assets';
-import { Navbar } from '@/components/navbar';
+import { Navbar } from '@/components/ui/navbar';
 import { ScrollToTopButton } from '@/components/scroll-to-top-button';
-import { Spacer } from '@/components/spacer';
+import { Spacer } from '@/components/ui/spacer';
 import Banner from '@/components/banner.astro';
 import Footer from '@/components/footer.astro';
 import MainLayout from '@/layouts/main.astro';
